requirements.txt
ament_copyright==0.9.8
ament_flake8==0.9.8
ament_pep257==0.9.8
cv_bridge==1.13.0.post0
ffmpeg==1.4
ffmpeg_python==0.2.0
matplotlib==3.7.5
numpy==1.24.4
openai==1.42.0
openai_whisper==20231117
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
Pillow==10.4.0
PyAudio==0.2.14
pykinect_azure==0.0.4
pynput==1.7.7
pytest==8.2.2
qi==3.1.5
Requests==2.32.3
segment_anything==1.0
setuptools==69.5.1
SpeechRecognition==3.10.3
supervision==0.22.0
torch==2.2.2
transformers==4.44.0
ultralytics==8.1.47
